+++ ecs/registry.go
@@ -3,22 +3,22 @@ package ecs
 import "reflect"
 
 // componentRegistry keeps track of component IDs
-type componentRegistry struct {
-	Components map[reflect.Type]ID
+type componentRegistry[T uint8] struct {
+	Components map[reflect.Type]T
 	Types      []reflect.Type
 }
 
 // newComponentRegistry creates a new ComponentRegistry
-func newComponentRegistry() componentRegistry {
-	return componentRegistry{
-		Components: map[reflect.Type]ID{},
+func newComponentRegistry[T uint8]() componentRegistry[T] {
+	return componentRegistry[T]{
+		Components: map[reflect.Type]T{},
 		Types:      make([]reflect.Type, MaskTotalBits),
 	}
 }
 
 // RegisterComponent registers a components and assigns an ID for it
-func (r *componentRegistry) RegisterComponent(tp reflect.Type, totalBits int) ID {
-	id := ID(len(r.Components))
+func (r *componentRegistry[T]) RegisterComponent(tp reflect.Type, totalBits int) T {
+	id := T(len(r.Components))
 	if int(id) >= totalBits {
 		panic("maximum of 128 component types exceeded")
 	}
@@ -28,7 +28,7 @@ func (r *componentRegistry) RegisterComponent(tp reflect.Type, totalBits int) ID
 }
 
 // ComponentID returns the ID for a component type, and registers it if not already registered
-func (r *componentRegistry) ComponentID(tp reflect.Type) ID {
+func (r *componentRegistry[T]) ComponentID(tp reflect.Type) T {
 	if id, ok := r.Components[tp]; ok {
 		return id
 	}
@@ -36,6 +36,6 @@ func (r *componentRegistry) ComponentID(tp reflect.Type) ID {
 }
 
 // ComponentType returns the type of a component by ID
-func (r *componentRegistry) ComponentType(id ID) reflect.Type {
+func (r *componentRegistry[T]) ComponentType(id T) reflect.Type {
 	return r.Types[id]
 }